requirements.txt
anyio==4.9.0
certifi==2025.6.15
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
python-dateutil==2.9.0.post0
python-telegram-bot==22.2
six==1.17.0
sniffio==1.3.1
typing_extensions==4.14.0
